chore($security-config): relocate security-config
- move from config to security package
