+++ module-core/src/main/java/com/woomoolmarket/config/WebConfig.java
@@ -1,24 +1,11 @@
 package com.woomoolmarket.config;
 
-import com.navercorp.lucy.security.xss.servletfilter.XssEscapeServletFilter;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.core.Ordered;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
-
-  @Bean
-  public FilterRegistrationBean<XssEscapeServletFilter> filterRegistrationBean() {
-    FilterRegistrationBean<XssEscapeServletFilter> filterRegistration = new FilterRegistrationBean<>();
-    filterRegistration.addUrlPatterns("/*");
-    filterRegistration.setFilter(new XssEscapeServletFilter());
-    filterRegistration.setOrder(Ordered.HIGHEST_PRECEDENCE + 1);
-    return filterRegistration;
-  }
 
   @Override
   public void addResourceHandlers(ResourceHandlerRegistry registry) {
