+++ module-core/src/main/java/com/woomoolmarket/security/config/OAuth2Properties.java
@@ -1,4 +1,4 @@
-package com.woomoolmarket.config.security;
+package com.woomoolmarket.security.config;
 
 import java.util.Map;
 import lombok.Getter;
